RuleSet#each_declaration respects order
The order in which rules are declared in a CSS file is important:

    h1 {
      margin: 0;
      font-size: 24px;
    }

    h1 {
      margin-bottom: 12px;
    }

As written, the bottom margin for `h1` elements is `12px`; but if the
two declarations were switched, the bottom margin would be `0`.

Because the `CssParser::RuleSet @declarations` attribute is a Hash,
declaration order is ignored (for Ruby version < 1.9). Hence this
commit.